refactor(options): move desktop-only settings to a dedicated Desktop page
System Tray, Startup and Search Engine were Electron-only sections wedged into
the "Other" page behind an isElectron() gate. Move them into their own
#electronOnly "Desktop" options page so the desktop settings have a clear home
and are hidden from the nav on web/server (with a placeholder when reached
directly, mirroring the Spellcheck page).
