+++ apps/client/src/widgets/type_widgets/ContentWidget.tsx
@@ -11,6 +11,7 @@ import MultiFactorAuthenticationSettings from "./options/multi_factor_authentica
 import EtapiSettings from "./options/etapi";
 import BackupSettings from "./options/backup";
 import SyncOptions from "./options/sync";
+import DesktopSettings from "./options/desktop";
 import OtherSettings from "./options/other";
 import InternationalizationOptions from "./options/i18n";
 import AdvancedSettings from "./options/advanced";
@@ -20,7 +21,7 @@ import "./ContentWidget.css";
 import { t } from "../../services/i18n";
 import BackendLog from "./code/BackendLog";
 
-export type OptionPages = "_optionsAppearance" | "_optionsShortcuts" | "_optionsTextNotes" | "_optionsCodeNotes" | "_optionsMedia" | "_optionsSpellcheck" | "_optionsPassword" | "_optionsMFA" | "_optionsEtapi" | "_optionsBackup" | "_optionsSync" | "_optionsOther" | "_optionsLocalization" | "_optionsSecurity" | "_optionsAdvanced" | "_optionsLlm";
+export type OptionPages = "_optionsAppearance" | "_optionsShortcuts" | "_optionsTextNotes" | "_optionsCodeNotes" | "_optionsMedia" | "_optionsSpellcheck" | "_optionsPassword" | "_optionsMFA" | "_optionsEtapi" | "_optionsBackup" | "_optionsSync" | "_optionsDesktop" | "_optionsOther" | "_optionsLocalization" | "_optionsSecurity" | "_optionsAdvanced" | "_optionsLlm";
 
 const CONTENT_WIDGETS: Record<OptionPages | "_backendLog", (props: TypeWidgetProps) => JSX.Element> = {
     _optionsAppearance: AppearanceSettings,
@@ -34,6 +35,7 @@ const CONTENT_WIDGETS: Record<OptionPages | "_backendLog", (props: TypeWidgetPro
     _optionsEtapi: EtapiSettings,
     _optionsBackup: BackupSettings,
     _optionsSync: SyncOptions,
+    _optionsDesktop: DesktopSettings,
     _optionsOther: OtherSettings,
     _optionsLocalization: InternationalizationOptions,
     _optionsSecurity: SecuritySettings,
