+++ apps/server/src/assets/translations/en/server.json
@@ -324,6 +324,7 @@
     "code-notes-title": "Code Notes",
     "images-title": "Media",
     "spellcheck-title": "Spellcheck",
+    "desktop-title": "Desktop",
     "security-title": "Security",
     "password-title": "Password",
     "multi-factor-authentication-title": "MFA",
